WIP - converting to class based approach
