+++ index.js
@@ -1,7 +1,8 @@
 const inquirer = require('inquirer');
 const colors = require('colors');
 const fs = require('fs');
-const generateSVG = require('./lib/generateSVG');
+const {Circle, Square, Triangle} = require('./utils/shapes');
+const generateSVG = require('./utils/shapes');
 const validColors = require('./colors.json');
 
 const questions = [
@@ -60,6 +61,25 @@ function validateColor(input) {
     }
     
     return 'Must enter a color keyword (i.e. lightseagreen) or a valid 6 digit hexadecimal number (i.e. #deb887) - try again'.brightRed;
+}
+
+function generateSVG(answers) {
+    if (answers.shape === 'Circle') {
+        return `<svg version="1.1" width="300" height="200">
+<circle cx="150" cy="100" r="100" fill="${answers.shapeColor}"/>
+<text x="150" y="120" font-size="50" text-anchor="middle" fill="${answers.color}">${answers.text}</text>
+</svg>`;
+    } else if (answers.shape === 'Square') {
+        return `<svg version="1.1" width="300" height="200">
+<rect width="300" height="200" fill="${answers.shapeColor}"/>
+<text x="150" y="120" font-size="50" text-anchor="middle" fill="${answers.color}">${answers.text}</text>
+</svg>`;
+    } else if (answers.shape === 'Triangle') {
+        return `<svg version="1.1" width="300" height="200">
+<polygon points="150,20 30,180 270,180" fill="${answers.shapeColor}"/>
+<text x="150" y="150" font-size="50" text-anchor="middle" fill="${answers.color}">${answers.text}</text>
+</svg>`;
+    }
 }
 
 init();
